New: Better handling around unsatisfied link errors.
Primarily a developer feature to warn when a local build of the SDK is broken because the Rust libraries did not properly link, which happens occasionally and can be frustrating but should never be seen by an end user. In the event that a broken build does make it to a user, this would at least improve the experience.