New: Capture performance metrics for scanning.
The SDK now provides hooks for receiving a callback whenever a scan completes that includes information about the number of blocks scanned and duration. Use this info to report device metrics around scan speeds. This is in anticipation of greatly improving these numbers in the near future. Before that, we want to start collecting a baseline measurement.