Fix: Address tough bug on older devices that crashed during restore.
The root issue seemed to be, if a keyboard was shown for a focused EditTextView that was also disabled, older versions of android would crash inside the input manager which was impossible to catch with a try/catch block. The end result was major: some users could not restore their wallets. This work around addresses the problem by keeping the focused view enabled at all times. In theory this could result in some weird behavior but in practice, that's unlikely.